requirements.txt
streamlit
tensorflow
pillow
numpy
plotly
flask
flask-cors
